Replace interface and Fix book search convert function
