+++ src/lib/client/Domain/Entities/BookSearch.ts
@@ -11,7 +11,26 @@ export interface BookSearch {
 	readonly description?: string;
 }
 
-export const convertGAPIResponseToBookSearch = (response: books_v1.Schema$Volume) => {
+/**書誌検索の検索結果一覧 */
+export type BookSearchResultListType = books_v1.Schema$Volumes;
+/**書誌検索の検索結果 */
+export type BookSearchResultType<T> = T extends books_v1.Schema$Volumes
+	? books_v1.Schema$Volume
+	: never;
+
+export const isGAPIResultList = (obj: any): obj is books_v1.Schema$Volumes => {
+	return 'totalItems' in obj;
+};
+
+export const isGAPIResult = (obj: any): obj is books_v1.Schema$Volume => {
+	return 'volumeInfo' in obj;
+};
+
+export const convertResponseToBookSearch = <
+	ResultType extends BookSearchResultType<BookSearchResultListType>
+>(
+	response: ResultType
+) => {
 	const bookSearch: BookSearch = {
 		keyId: response.id!,
 		title: response.volumeInfo?.title,
